Changed the way the subgrid models are constructed and operate.
This was done to enable better re-use of the existing macro-scale simulations tools and to provide a cleaner implementation that is easier to understand and modify.

In some tests, the new approach is about 2x faster than the previous one and uses much less memory.

These changes required numerous small modifications throughout the code.