Added the ability to evaluate multiple subgrid models at the same time.  This improves performance by up to 4x on some problems if the workset size is chosen appropriately.  There is still room for improvement as this new capability has not been optimized.
Current constraint: the workset size must evenly divide the total number of elements.  This is not a restriction for the coarse scale model, but I believe the current fault is due to a zero block in the subgrid Jacobian if all of the spaces are not used.  Should be a simple fix (adding ones on the diagonal in this block) but have not tried it yet.

Also made a slight change to how the cells decide which subgrid model to use which required updating a couple of regression tests.

Due to an incompatibility, the older version of the subgrid model no longer works and will be deprecated.  The subgrid explicit model was a copy of this older subgrid model and will be removed.  It was not using the explicit time integrators yet anyway.  The corresponding regression test has been de-activated.