Minor change that forced me to modify numerous files.  Changed the function manager to be sinlge block, which removes a std::vector, but requires a vector of function managers instead of a single one.
Since each function manager can only be associated with one workset, this was a mistake to have a function manager handling multiple blocks.  This is probably why the older multi-block test does not pass.

Removing this also reduced some arguments in the set-up of the function manager, so the physics modules no longer need to be aware of the blocknum.