Completely revampled the time integration capabilities.
Time integration methods now require a BDF order (default is 1) and a Butcher tableau (default is 'BWE').

The Butcher Tableau key is a string that points to the Butcher A,b,c implemented in the solver interface.

If the BDF order is higher than 1, then a different startup procedure is required to achieve the optimal order.

Also added a simple ODE solver for testing.

Most of the capabilities are tested in the new sets of regression tests.