Rather large change with respect to strongly imposed Dirichlet conditions.  This capability was inherited from a previous version of the code and was really expensive and limited to nodal HGRAD discretizations.
The new capability uses an L2-projection (instead of an interpolant) to construct a representation of the Dirichlet data in the discretization space.  This is how the initial conditions were generalized.  This works for all basis types.

The drawback is that L2-projections are not free, so there are two cheaper options.

Cheaper option 1 (cheapest) - each variable on each block uses scalar Dirichlet data.  This actually covers most use cases with zero DBCs.  This avoids the L2-projection entirely.  This concept was also extended to the initial conditions.

Cheaper option 2 - non-transient Dirichlet data.  In many cases, the Dirichlet data does not change over time, so the L2-projection only needs to occur once.  This is the default option.

Also changed how the workset computes the solution at the volumetric integration points.  This is slightly more expensive, but reduces almost 1000 lines of redundant code and can be re-used for side and face values (not implemented yet).

Due to these changes, a few regression tests had to be re-based.