Adding a basis ability to plot the cell-average of HFACE variables.
The volumetric average of an HFACE variable is zero, so we compute the face/edge average and plot it as piecewise constant data over the cell.

Even using in-situ visualization, this requires some extra effort because the solutions on the faces are not always computed.

So, the default is for this option to be turned off.  If visualization is enabled and HFACE variables are present, then the user will receive a warning and instructions for turning this on if so desired.