Made a couple of changes to the CMake configure script.
Changed how the cells, worksets and physics modules store and use integration information and basis functions.

DynRankViews are now only used when necessary and rarely saved.  Instead, a Kokkos::View of the appropriate size and dimensions is filled and stored.

This does require a few extra deep copies, but this is minimal (on the order of the number of cells).

The benefits are robustness and avoidance of UVM.

The DRVs now use PHX::Device to be consistent with the rest of Trilinos on UVM-enabled devices.  However, MrHyDE does not require UVM otherwise.

Robustness comes via how data is accessed in a View versus a DRV.  DRVs can accept access indices that are different from their rank.  Views cannot.  Fixed several issues associated with this.