Removed DRVs from one of the physics modules.
Removed DRVs from worksets.  This also includes a minor change in the constructor.  Some DRVs were being passed in just for the number of integration points.  This can be obtained from the cell meta-data.

Moved the cellinfo vector out of the cell and into the assembly device.  This mostly used information from the cell meta data anyways.

Made a minor change in the workset for the residual view.  The size of this is now based on maxDerivs instead of the number of DOF per element.  This should be more robust and allow us to completely avoid resizing this View.  I suppose there could have been a problem if the discretized parameters required more DOFs on an element than the state variables.  This fix should cover this case.  The array is slightly larger if maxDerivs is set larger than necessary, but I don't think this causes an efficiency issue due to how the parallel_for ranges are set.