Made quite a few changes to generalize where certain computations can be performed.  The linear algebra objects are no longer confined to the HostDevice and can be assigned to a GPU if this capability is enabled.  This caused several interfaces to need to be templated for the multiscale cases where the global linear algebra may use the CPUs and the local fine scale linear algebra may use the GPUs.
The implementation should be quite general and allow for the LA/Assembly to be CPU/CPU, GPU/GPU, CPU/GPU and GPU/CPU.  This will become more complicated when the subgrid LA and subgrid solver options are added.

This has not been fully tested, but all of the expected tests pass if all devices are the HostDevice.