Found an oversight in the flux evaluation with large worksets.
The outer-most loop was using the size of the normals array, but this was recently changed to be stationary to enable the function manager to access the normals.  As a result, this array may be tto large for the boundary cells for large worksets.  A temporary fix has been added to use the mesh size array instead.  Other loops in the physics modules are not affected because they use the basis, which is always the correct size.