Major change to how the worksets store and utilize Kokkos Views.
We need to use views that use a contiguous layout and we cannot use subviews of these views.  Previously, the data was stored as several rank-4 views and subviews were used as needed.  This required additional objects for each new data type, e.g., the aux variables.  Now, all of this data is stored as a rank-2 view and these views are stored in a vector of views (views of views are fragile).  Data is accessed by calling either getData(), getDataSc(), or just get().  Each of these have different arguments.  Because the workset searches over a list of strings, it may be necessary to avoid repeated searches if the same data is accessed multiple times.  The thermal physics module shows how to do this.

All physics modules have been updated to the new format, but only thermal avoids the repeated searches (for now).

The workset has been substantially changed and many functions have been removed or combined.  The new version is much simpler.

The function manager has been updated to properly access the data in the workset.  The new version is also much simpler.

The only known remaining issue is associated with optimization using sensors.  This is causing a few tests to fail, but the rest should pass.