Making it easier to add new physics modules.
The physics interface, which will soon be moved to a manager, no longer needs to be modified.

The physicsImporter, which is in src/physics is the only file that needs to be touched and these modifications are trivial.

Also added the ability to define extra variables.  This is not tested yet.

Also cleaned up a couple of regression test input files.