Apparently, the Extra variables work exactly as intended.
These can be used for multiscale simulations if one does not want to define physics modules on the coarse scale.  These were not being used anyways.

This regression test shows how to do this.

Note: defining extra variables that do not have associated equations will result in a large block of zeros in the Jacobian matrix.  This will probably cause an error to occur.