Finished creating the "default" input file.
These files contain all of the options in MrHyDE that are available to the user through the input file (unless I missed some, which is likely).

These are just for reference and should not be used for creating new examples/tests.

Also removed some unused options from several files.

Corrected an error in the transient multiscale capability.  It was not reading the final time in properly.  Rebased the corresponding test.