Re-organized a few parts of the code.
Changed the analysis interface to a manager since it doesn't really interface with any Trilinos or external (to MrHyDE) packages.

Split the solver interface into a linear algebra interface, which connects to the Trilinos vectors, matrices, linear solvers and preconditioners, and a solver manager, which handles MrHyDE's solution procedures.

Also added the linear algebra objects associated with the auxiliary variables and discretized parameters.