Making this regression test inactive.
This tests the optimization gradient with the multiscale models.  Due to a complete refactor of the optimization capabilities, this also needs to be rewritten, but I don't have time for it right now.

This test will remain in the repository and can be re-activated any time in the mrhyde.tst file.