Added a compile time option for changing the vector size in the hierarchical parallelism.  This can improve performance on a GPU and should not affect anything on the CPU.
The default is set to the number of derivatives, but this can be adjusted from the configure file.

Adjusted the files that previously had the vector size hard coded to 32.  This was less places than I thought.  This means that most of the physics modules are not optimal for GPU runs.  The thermal module is really the only one.  This requires a simple change to the parallel_for loops.

Also added some better configuration output.