After a long search for a memory leak, it turns out there is no memory leak.
However, it also turns out that using certain Kokkos profiling tools can results in very large output files if the code uses lots of timers.  This took quite a while to find.  Leaving a few timers in the function manager commented out for now.

Most of these changes were to isolate the fake memory leak issue and are very minor.