Switched all of the physics modules to use vistas instead of views when using the function manager.
A few managers/interfaces had to be updated as well.

Overview of vistas:

A vista is basically a light wrapper for a 2-dimensional view.

It is designed for cases where only the operator () is needed and where data might not need to be stored as AD objects.

The function manager is careful about how it stores data.  For example, if all of the leafs on a given branch are constants, then the branch data is stored as a constant.
However, by requiring the function manager to return a View_AD2 of the appropriate size when calling evaluate(), this constant value was being copied into a large view of AD objects, which is wasteful.  Vistas make this unnecessary.

Vistas can be created with different types of data and the operator () returns the appriate value based on the storage type.  While this was designed to improve memory usage, preliminary tests indicate up to 40% reduction in data access/usage.

I initially added an extent function to the vistas, but commented it out because this could cause some strange behavior if the vista only stores scalar values.