Adding a regression test for the new fully explicit solver.
This solver only works with explicit time integration methods (obviously) and with lumped mass matrices.  No Jacobians are used in the solves.

However, the Jacobians are still being calculated since AD is still used.

Turning off the AD is now possible, but only for these explicit solves.  This requires a compile flag to switch the AD type from Sacado::SFAD to ScalarT (double), so it cannot be tested in this regression test.