Changed some of the names associated with the porous media modules.
HDIV -> Mixed
HDIV_HYBRIDIZED -> Mixed_Hybridized
HDIV_weakGalerkin -> WeakGalerkin

Also changed the names of the regression tests.

Modified the KL expansion tool to use Kokkos Views.  Also changed how the nonlinear solver finds the roots in the characteristic equation.  This is necessary is a large number of roots are needed.

Changed the mixed porous formulation to use a KL expansion for the permeability.  This is only implemented and tested in 1D.  Extending to 2D and 3D should be easy.

Added a appending string to the visualization quantities.  This is used by the UQ analyis mode to visualize all realizations (use carefully!).