Fixed a typo in the Dirichlet BCs for HDIV variables in 2D.
Added a 2D and a 3D KL expansion within the porousMixed module.  Most of this code does not belong here, but I eventually need to rewrite the KL expansion tool to be multi-dimensional.  This is the first time the order of the terms in the expansion really mattered.

Also adding regression tests for 2D and 3D.