Make the Kuramoto-Sivashinsky model run without crashing and add a test.
I was using the antiquated version of grabbing offsets, which was resulting in the program being stuck in getOffsets(w) indefinitely, which is why the w equation was not working as intended. I haven't run this on any benchmarks or eyeballed the results yet, but this should already be much better than it was before.