Made a rather substantial change in how state/aux/discretized parameters are computed at quadrature points.
Previously, this was performed for all available fields, e.g., solution, grad, time derivative, etc., for all stages of the time integration before the physics modules computed the residuals.

For certain methods, which don't use all of the available information for each stage, this performs excessive computations and uses extra memory.

The new capability computes these on-demand.  The workset does not compute a field at the quadrature points until it is needed.  There is an update flag that makes sure this computation only happens once per field per stage.

This computation gets performed whenever the getData function is called and the data has not been computed yet.  The thermal physics module used a slightly different mechanism to access this data and had to be updated to use this function.

The previous capability is still functional and can be wnabled by changing the onDemand flag in the workset.