Hface initialcondition (#3)
* New branch for issue :: enabling projection of IC for HFACE variables.

* Sketch of what I want to do here.

There are routines for projecting/handling Dirichlet data, so I want to
expand that functionality to handle IC projection as well.

I'm not sure the IC gets stored by the function manager (at side IP)
ever, though.

One thing to consider is if this will conflict with BC enforcement...
but this should only happen once.

* A halfway decent stab at making the plan happen, but not sure this is
going to work out.

One particular issue is that the L2 projector is never actually seeing
the interface "boundaries". So although HFACE vars are defined only on
the side ip, I don't seem to be accessing them currently.

* Rolling back after a few discussions with Tim, and a new implementation.

The functionality is exists in cell, not boundary cell.
No segfaulting, but don't have a way right now to check that the
projection is working.

* Automatically protect against HFACE preconditioner issues for IC
projection.
I don't see a great way to do the same for BC issues, so that will
be deferred to the user for now.

* New branch for issue :: enabling projection of IC for HFACE variables.

* Sketch of what I want to do here.

There are routines for projecting/handling Dirichlet data, so I want to
expand that functionality to handle IC projection as well.

I'm not sure the IC gets stored by the function manager (at side IP)
ever, though.

One thing to consider is if this will conflict with BC enforcement...
but this should only happen once.

* A halfway decent stab at making the plan happen, but not sure this is
going to work out.

One particular issue is that the L2 projector is never actually seeing
the interface "boundaries". So although HFACE vars are defined only on
the side ip, I don't seem to be accessing them currently.

* Rolling back after a few discussions with Tim, and a new implementation.

The functionality is exists in cell, not boundary cell.
No segfaulting, but don't have a way right now to check that the
projection is working.

* Automatically protect against HFACE preconditioner issues for IC
projection.
I don't see a great way to do the same for BC issues, so that will
be deferred to the user for now.

* Bring master changes in. Update a re gression test to cover the new functionality.

Co-authored-by: Bryan Reuter <bwreute@sandia.gov>