Made a minor change to the parameter list that gets passed into the physics modules.
Previously, this was all of the settings.

Now, it is just the physics settings associated with the appropriate block.

Also, leveraged the flexibility in the Vistas to allow functions to be created with scalars (doubles) instead of just with strings.

Finally, updated the mirage physics to use the mirage planewave source and PMLs.