Add the physicsTest module and a corresponding test.
The physicsTest model is a procedural physics model that generates and solves a simple L^2 projection or Laplace problem, and currently outputs the basis function values or gradient values at quadrature points. This means the gold file is essentially a plot of the basis functions, which helps to diagnose issues while also serving as a utility to improve additional and more complex discretizations. This is intended to support a wide array of configurations, but currently supports HGRAD (projection, Laplacian), HDIV (projection), HDIV_AC (projection), HCURL (projection) configurations.

There is also an option to rename variables, but currently only the variable name p works. I may remove it later if it seems there are no potental merits in that feature.