Made a few minor changes to how elements are grouped into cells.
The multiscale runs now automatically use a specialized grouping scheme that avoids grouping elements that have different boundary faces into one cell.  The was causing an issue in the subgrid models.

Also removed an assumption that the number of elements in the first cell is largest and always larger than or equal to the number of elements in the first boundary cell.

Finally, updated one regression test that changed slightly due to different element groupings and how subgrid models are chosen.