Exposing a few more options within the mirage translator.
Fixed an issue with building the mass matrices with maxDerivs set very low.  This was using the same allocation for the scatter operation which used maxDerivs to set the max number of entries to add at a time.  This makes sense for implicit methods, but not explicit.  Hard-coded to 64 for now, but this should cover most use cases.