A rather large commit that affects most of the infrastructure, but not the physics modules.
MrHyDE has always been able to solve coupled multiphysics problems, but only in a fully-coupled monolithic manner.

This commit enables the use of multiple physics sets, which can be used to enable loose, or iterative coupling, as well as certain operator splitting schemes.

While I was modifying the infrastructure of the code, I also made a few changes to enhance robustness and performance.