Simplified the constructors for the solution fields and the scalar fields.  This will make it much easier to add new fields from outside the workset.
The new format infers the derivatives and the location of the evaluations from the variable name.  This means that variables cannot have "_t" in their names.  This required a change in one of the regression tests.

Also added new getSolutionField and getScalarField in workset.  These will eventually replace getData and getDataSc.