Simplifying the syntax for storing the basis functions.  Now, only the storage proportion is required.
In doing so, decided to comment out a fair bit of code that was difficult to maintain and really only existed to reduce the number of memory allocations on the GPU.  While this helped a little, the benefits were not significant enough to justify the additional memory usage.

For now, all of this code is commented out.  Once I am sure that there are no use cases that really require this, I will delete the commented sections.