This commit includes the ability to add solution fields and scalar fields to the workset from outside the workset.
The functions that enable this are also now called in the workset when the functions are created.

Updated the test_functions sandbox file to use this functionality, which is much simpler than before.

Updated the interpreter and function manager to handle <= and >=.  It was missing the logic required for 2 character operators.

Updated the interpreter and function manager to consider pointwise min and max.  Previously, min(a) and max(a) referred to the min and max (respectively) of a field (a) over the quadrature points within a face or element.  This functionality is now accessed using emin and emax.  Now, we use min(a,b) and max(a,b) to compute the pointwise min and max (respectively) between two arrays (a and b).