Created a simplified version of the subgrid DtN maps where the subgrid models are unaware of the macro-scale BCs.  They only produce DtN maps.
While this does not add anything new, it is beneficial for a few reasons:

1.  More uniformity in the subgrid models - they all do the same thing.
2.  Specialized assembly partitioning schemes are not required.
3.  No more hanging edges/faces on Neumann/flux boundaries.
4.  Allows for finally getting rid of the sideinfo arrays, which are large and only used to communicate coarse scale BCs to the subgrid models.

The drawback in using this new version is that the boundary conditions are only enforced at the coarse scale, which means that fine scale effects cannot be captured.  However, there are currently no use cases this.

In order to enforce coarse scale flux conditions, added a new Flux boundary condition.  This mostly duplicates the Neumann BC capability, but does not require a physics module and therefore can also be used with extra variables.