A large commit, but mostly just renaming objects and variables.
The main change is removing the cells and boundary cells.

In the initial versions of the code, cell was defined similarly to how Intrepid defines cell, e.g., an element in the mesh.

With the introduction of Kokkos and on-node parallelism, cells were overloaded to mean a collection of elements.  This was always meant to be replaced, but the cells persisted for quite a while.

Now, we have renamed cell -> group and boundaryCell -> boundaryGroup.  This better reflects that these objects store the information for a group of elements.

Also made an effort to be more explicit in some of the loops over block and groups, e.g., b->block and e->grp.

Also made sure each compiler definition includes a MRHYDE prefix to avoid conflicts if we want to use MrHyDE with other packages.  I am not aware of any such conflicts, but these can be difficult to diagnose.

Finally, changed usernum in the subgrid models to macrogrp.