Correct the use of xml minidom document writing inside runtests.py
python3 runtests.py no longer crashes on a system that contains both python2 and python3. With that said, the subprocesses technically run commands like './mrhyde.tst -n 4', which likely means they are deferring to the operating system to choose and interpreter for the file.
In this case, the shebang present at the start of each mrhyde.tst is a python2 invocation, so these shebangs likely need to be changed en-masse as well if a machine does not support python2.