Adding a finer grained control over the storage of the integration data and the basis functions.
As of this commit, the groups and boundary groups always store the integration points, the integration weights and the orientations.  Storage of the basis functions is optional.

Added a new option to create a database on each processor/block of basis data that can be reused if the element jacobians and orientations match.  This can lead to a >99 percent reduction in the storage cost for uniform meshes and >90 percent savings for certain other meshes without significantly increasing the run time.

This capability is not enabled by default and should be considered extremely experimental.