Remove the opts.baseline part of each mrhyde.tst and add IOSS to the logfile cleaner
opts.baseline was part of a set of features that was never really utilized. many of the other opts checks are still available in the mrhyde.tst files. additionally, this change will leave some awkward empty lines in each of the mrhyde.tst files around line 54 in each file, but since it's not the same line in every file, I elected to avoid trying to remove the awkward whitespace while I'm doing batch editing.