Finish the regression testing update to support both python2 and python3. This closes issue #8
The regression test suite should now be completely agnostic to the version of python that is used. The only bit of code required to maintain both python2 and python3 is the use of the __future__ module in mrhyde_test_support.py to make the pretty verbose line printing work in both python2 and python3.