Made a fairly significant change that affects the worksets, physics modules and a few other places.
Previously, field on a side or at a point required including " side" or " point" to access this data from the workset.

This is now handled at a higher level by MrHyDE.  Thus, writing getSolutionField("u") will "automatically" return the proper fields.

This is managed through flags in the wkset (isOnSide and isOnPoint).  The default for both of these is false.

If one of these needs to be turned on, e.g., for a boundary residual, this should happen at the highest level possible and reset afterwards.

Also, while changing the syntax in the physics modules, changed how the normals and tangents are accessed.  Instead of "nx", we now use "n[x]". This is more consistent with other vector quantities.

All physics modules and regression test input files have been updated to use the new syntax.

If something breaks, then it probably isn't covered by a test.