Significant change to the basis database usage.
Previously, it would copy the database basis into another view to be used by the worksets and physics modules.  This was a waste of time and memory.

Now, the code always stores a basis_index, which is just 0 through numElem if the database is not used.  Otherwise, it is the database index.

The workset now uses this index to map the DOFs to the solutions at integration points.

The change is semi-backwards compatible for the physics modules.  If a module is not going to be used along with the database, then no change is necessary.

If the module might be used with the database, then a simple change is required in each parallel_for loop that involves a basis.

Examples of the change are in burgers.cpp and cdr.cpp, which is how far I got until I realized the change was optional.

Eventually, all physics modules will incorporate this change.