Re-enabled the dynamic subgrid modeling for adaptive multiscale modeling (needed for WCCM talk).
MrHyDE has always been able to handle different subgrid models in different parts of the computational domain,
but switching them adaptively has been disabled for a little while.

This was disabled for two reasons:

1. The switch to Tpetra broke the projections between subgrid state solutions.
2. The error calculations (for verification) were not able to recognize when certain subgrids were no longer is use.

Both of these issues are fixed with this commit and the dynamic multiscale test has been updated.