Changed DatabaseView to CompressedView.
The DatabaseView wasn't actually used anywhere, but now the CompressedView is the default storage mechanism for basis functions and local mass matrices.

By using this new object, the physics modules can index into the basis functions as they normally would, and the view automatically returns the proper data.

There are still a few parts of the code that don't use the data compression, e.g., the face calculations, but now changes are not required in the physics modules to use compression.

This should simplify generalizing the data compression later on.