Run a python script to generate template physics documentations.
This autopopulates some simple doxygen comments on every physics class with input deck options and a placeholder LaTeX equation to describe the volume residual. The goal of this is to start some documentation and overcome the energy barrier so that anybody visiting the file can make minor modifications to fix the details that are missing.